+++ CHANGELOG.md
@@ -5,6 +5,12 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `InputType::Group`, `InputItem::Group`, and `ListItem::Group`
+
 ## [0.2.0] - 2022-11-27
 
 ### Changed
@@ -24,6 +30,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Initial release.
 
+[Unreleased]: https://github.com/aicers/frontary/compare/0.2.0...main
 [0.2.0]: https://github.com/aicers/frontary/compare/0.1.1...0.2.0
 [0.1.1]: https://github.com/aicers/frontary/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/aicers/frontary/tree/0.1.0